requirements.txt
langdetect
PyGithub
python-dotenv
celery
elasticsearch>=2.0.0,<3.0.0
redis